Crate spl_token

An ERC20-like Token program for the Solana blockchain

Re-exports

pub use solana_program;

Modules

error

Error types

instruction

Instruction types

native_mint

The Mint that represents the native token

processor

Program state processor

state

State transition types

Statics

ID

The static program ID

Functions

amount_to_ui_amount

Convert a raw amount to its UI representation (using the decimals field defined in its mint)

amount_to_ui_amount_string

Convert a raw amount to its UI representation (using the decimals field defined in its mint)

amount_to_ui_amount_string_trimmed

Convert a raw amount to its UI representation using the given decimals field Excess zeroes or unneeded decimal point are trimmed.

check_id

Confirms that a given pubkey is equivalent to the program ID

check_program_account

Checks that the supplied program ID is the correct one for SPL-token

id

Returns the program ID

try_ui_amount_into_amount

Try to convert a UI represenation of a token amount to its raw amount using the given decimals field

ui_amount_to_amount

Convert the UI representation of a token amount (using the decimals field defined in its mint) to the raw amount