Enum spl_token::instruction::TokenInstruction

#[repr(C)]
pub enum TokenInstruction<'a> {

    InitializeMint {
        decimals: u8,
        mint_authority: Pubkey,
        freeze_authority: COption<Pubkey>,
    },
    InitializeAccount,
    InitializeMultisig {
        m: u8,
    },
    Transfer {
        amount: u64,
    },
    Approve {
        amount: u64,
    },
    Revoke,
    SetAuthority {
        authority_type: AuthorityType,
        new_authority: COption<Pubkey>,
    },
    MintTo {
        amount: u64,
    },
    Burn {
        amount: u64,
    },
    CloseAccount,
    FreezeAccount,
    ThawAccount,
    TransferChecked {
        amount: u64,
        decimals: u8,
    },
    ApproveChecked {
        amount: u64,
        decimals: u8,
    },
    MintToChecked {
        amount: u64,
        decimals: u8,
    },
    BurnChecked {
        amount: u64,
        decimals: u8,
    },
    InitializeAccount2 {
        owner: Pubkey,
    },
    SyncNative,
    InitializeAccount3 {
        owner: Pubkey,
    },
    InitializeMultisig2 {
        m: u8,
    },
    InitializeMint2 {
        decimals: u8,
        mint_authority: Pubkey,
        freeze_authority: COption<Pubkey>,
    },
    GetAccountDataSize,
    InitializeImmutableOwner,
    AmountToUiAmount {
        amount: u64,
    },
    UiAmountToAmount {
        ui_amount: &'a str,
    },

}

Instructions supported by the token program.

Variants

InitializeMint

Fields

decimals: u8

Number of base 10 digits to the right of the decimal place.

mint_authority: Pubkey

The authority/multisignature to mint tokens.

freeze_authority: COption<Pubkey>

The freeze authority/multisignature of the mint.

Initializes a new mint and optionally deposits all the newly minted tokens in an account.

The InitializeMint instruction requires no signers and MUST be included within the same Transaction as the system program’s CreateAccount instruction that creates the account being initialized. Otherwise another party can acquire ownership of the uninitialized account.

Accounts expected by this instruction:

[writable] The mint to initialize.

[] Rent sysvar

InitializeAccount

Initializes a new account to hold tokens. If this account is associated with the native mint then the token balance of the initialized account will be equal to the amount of SOL in the account. If this account is associated with another mint, that mint must be initialized before this command can succeed.

The InitializeAccount instruction requires no signers and MUST be included within the same Transaction as the system program’s CreateAccount instruction that creates the account being initialized. Otherwise another party can acquire ownership of the uninitialized account.

Accounts expected by this instruction:

[writable] The account to initialize.

[] The mint this account will be associated with.

[] The new account’s owner/multisignature.

[] Rent sysvar

InitializeMultisig

Fields

m: u8

The number of signers (M) required to validate this multisignature account.

Initializes a multisignature account with N provided signers.

Multisignature accounts can used in place of any single owner/delegate accounts in any token instruction that require an owner/delegate to be present. The variant field represents the number of signers (M) required to validate this multisignature account.

The InitializeMultisig instruction requires no signers and MUST be included within the same Transaction as the system program’s CreateAccount instruction that creates the account being initialized. Otherwise another party can acquire ownership of the uninitialized account.

Accounts expected by this instruction:

[writable] The multisignature account to initialize.

[] Rent sysvar

..2+N. [] The signer accounts, must equal to N where 1 <= N <= 11.

Transfer

Fields

amount: u64

The amount of tokens to transfer.

Transfers tokens from one account to another either directly or via a delegate. If this account is associated with the native mint then equal amounts of SOL and Tokens will be transferred to the destination account.

Accounts expected by this instruction:

Single owner/delegate

[writable] The source account.

[writable] The destination account.

[signer] The source account’s owner/delegate.

Multisignature owner/delegate

[writable] The source account.

[writable] The destination account.

[] The source account’s multisignature owner/delegate.

..3+M [signer] M signer accounts.

Approve

Fields

amount: u64

The amount of tokens the delegate is approved for.

Approves a delegate. A delegate is given the authority over tokens on behalf of the source account’s owner.

Accounts expected by this instruction:

Single owner

[writable] The source account.

[] The delegate.

[signer] The source account owner.

Multisignature owner

[writable] The source account.

[] The delegate.

[] The source account’s multisignature owner.

..3+M [signer] M signer accounts

Revoke

Revokes the delegate’s authority.

Accounts expected by this instruction:

Single owner

[writable] The source account.

[signer] The source account owner.

Multisignature owner

[writable] The source account.

[] The source account’s multisignature owner.

..2+M [signer] M signer accounts

SetAuthority

Fields

authority_type: AuthorityType

The type of authority to update.

new_authority: COption<Pubkey>

The new authority

Sets a new authority of a mint or account.

Accounts expected by this instruction:

Single authority

[writable] The mint or account to change the authority of.

[signer] The current authority of the mint or account.

Multisignature authority

[writable] The mint or account to change the authority of.

[] The mint’s or account’s current multisignature authority.

..2+M [signer] M signer accounts

MintTo

Fields

amount: u64

The amount of new tokens to mint.

Mints new tokens to an account. The native mint does not support minting.

Accounts expected by this instruction:

Single authority

[writable] The mint.

[writable] The account to mint tokens to.

[signer] The mint’s minting authority.

Multisignature authority

[writable] The mint.

[writable] The account to mint tokens to.

[] The mint’s multisignature mint-tokens authority.

..3+M [signer] M signer accounts.

Burn

Fields

amount: u64

The amount of tokens to burn.

Burns tokens by removing them from an account. Burn does not support accounts associated with the native mint, use CloseAccount instead.

Accounts expected by this instruction:

Single owner/delegate

[writable] The account to burn from.

[writable] The token mint.

[signer] The account’s owner/delegate.

Multisignature owner/delegate

[writable] The account to burn from.

[writable] The token mint.

[] The account’s multisignature owner/delegate.

..3+M [signer] M signer accounts.

CloseAccount

Close an account by transferring all its SOL to the destination account. Non-native accounts may only be closed if its token amount is zero.

Accounts expected by this instruction:

Single owner

[writable] The account to close.

[writable] The destination account.

[signer] The account’s owner.

Multisignature owner

[writable] The account to close.

[writable] The destination account.

[] The account’s multisignature owner.

..3+M [signer] M signer accounts.

FreezeAccount

Freeze an Initialized account using the Mint’s freeze_authority (if set).

Accounts expected by this instruction:

Single owner

[writable] The account to freeze.

[] The token mint.

[signer] The mint freeze authority.

Multisignature owner

[writable] The account to freeze.

[] The token mint.

[] The mint’s multisignature freeze authority.

..3+M [signer] M signer accounts.

ThawAccount

Thaw a Frozen account using the Mint’s freeze_authority (if set).

Accounts expected by this instruction:

Single owner

[writable] The account to freeze.

[] The token mint.

[signer] The mint freeze authority.

Multisignature owner

[writable] The account to freeze.

[] The token mint.

[] The mint’s multisignature freeze authority.

..3+M [signer] M signer accounts.

TransferChecked

Fields

amount: u64

The amount of tokens to transfer.

decimals: u8

Expected number of base 10 digits to the right of the decimal place.

Transfers tokens from one account to another either directly or via a delegate. If this account is associated with the native mint then equal amounts of SOL and Tokens will be transferred to the destination account.

This instruction differs from Transfer in that the token mint and decimals value is checked by the caller. This may be useful when creating transactions offline or within a hardware wallet.

Accounts expected by this instruction:

Single owner/delegate

[writable] The source account.

[] The token mint.

[writable] The destination account.

[signer] The source account’s owner/delegate.

Multisignature owner/delegate

[writable] The source account.

[] The token mint.

[writable] The destination account.

[] The source account’s multisignature owner/delegate.

..4+M [signer] M signer accounts.

ApproveChecked

Fields

amount: u64

The amount of tokens the delegate is approved for.

decimals: u8

Expected number of base 10 digits to the right of the decimal place.

Approves a delegate. A delegate is given the authority over tokens on behalf of the source account’s owner.

This instruction differs from Approve in that the token mint and decimals value is checked by the caller. This may be useful when creating transactions offline or within a hardware wallet.

Accounts expected by this instruction:

Single owner

[writable] The source account.

[] The token mint.

[] The delegate.

[signer] The source account owner.

Multisignature owner

[writable] The source account.

[] The token mint.

[] The delegate.

[] The source account’s multisignature owner.

..4+M [signer] M signer accounts

MintToChecked

Fields

amount: u64

The amount of new tokens to mint.

decimals: u8

Expected number of base 10 digits to the right of the decimal place.

Mints new tokens to an account. The native mint does not support minting.

This instruction differs from MintTo in that the decimals value is checked by the caller. This may be useful when creating transactions offline or within a hardware wallet.

Accounts expected by this instruction:

Single authority

[writable] The mint.

[writable] The account to mint tokens to.

[signer] The mint’s minting authority.

Multisignature authority

[writable] The mint.

[writable] The account to mint tokens to.

[] The mint’s multisignature mint-tokens authority.

..3+M [signer] M signer accounts.

BurnChecked

Fields

amount: u64

The amount of tokens to burn.

decimals: u8

Expected number of base 10 digits to the right of the decimal place.

Burns tokens by removing them from an account. BurnChecked does not support accounts associated with the native mint, use CloseAccount instead.

This instruction differs from Burn in that the decimals value is checked by the caller. This may be useful when creating transactions offline or within a hardware wallet.

Accounts expected by this instruction:

Single owner/delegate

[writable] The account to burn from.

[writable] The token mint.

[signer] The account’s owner/delegate.

Multisignature owner/delegate

[writable] The account to burn from.

[writable] The token mint.

[] The account’s multisignature owner/delegate.

..3+M [signer] M signer accounts.

InitializeAccount2

Fields

owner: Pubkey

The new account’s owner/multisignature.

Like InitializeAccount, but the owner pubkey is passed via instruction data rather than the accounts list. This variant may be preferable when using Cross Program Invocation from an instruction that does not need the owner’s AccountInfo otherwise.

Accounts expected by this instruction:

[writable] The account to initialize.

[] The mint this account will be associated with.

[] Rent sysvar

SyncNative

Given a wrapped / native token account (a token account containing SOL) updates its amount field based on the account’s underlying lamports. This is useful if a non-wrapped SOL account uses system_instruction::transfer to move lamports to a wrapped token account, and needs to have its token amount field updated.

Accounts expected by this instruction:

[writable] The native token account to sync with its underlying lamports.

InitializeAccount3

Fields

owner: Pubkey

The new account’s owner/multisignature.

Like InitializeAccount2, but does not require the Rent sysvar to be provided

Accounts expected by this instruction:

[writable] The account to initialize.

[] The mint this account will be associated with.

InitializeMultisig2

Fields

m: u8

The number of signers (M) required to validate this multisignature account.

Like InitializeMultisig, but does not require the Rent sysvar to be provided

Accounts expected by this instruction:

[writable] The multisignature account to initialize.

..1+N. [] The signer accounts, must equal to N where 1 <= N <= 11.

InitializeMint2

Fields

decimals: u8

Number of base 10 digits to the right of the decimal place.

mint_authority: Pubkey

The authority/multisignature to mint tokens.

freeze_authority: COption<Pubkey>

The freeze authority/multisignature of the mint.

Like InitializeMint, but does not require the Rent sysvar to be provided

Accounts expected by this instruction:

[writable] The mint to initialize.

GetAccountDataSize

Gets the required size of an account for the given mint as a little-endian u64.

Return data can be fetched using sol_get_return_data and deserializing the return data as a little-endian u64.

Accounts expected by this instruction:

[] The mint to calculate for

InitializeImmutableOwner

Initialize the Immutable Owner extension for the given token account

Fails if the account has already been initialized, so must be called before InitializeAccount.

No-ops in this version of the program, but is included for compatibility with the Associated Token Account program.

Accounts expected by this instruction:

[writable] The account to initialize.

Data expected by this instruction: None

AmountToUiAmount

Fields

amount: u64

The amount of tokens to reformat.

Convert an Amount of tokens to a UiAmount string, using the given mint. In this version of the program, the mint can only specify the number of decimals.

Fails on an invalid mint.

Return data can be fetched using sol_get_return_data and deserialized with String::from_utf8.

Accounts expected by this instruction:

[] The mint to calculate for

UiAmountToAmount

Fields

ui_amount: &'a str

The ui_amount of tokens to reformat.

Convert a UiAmount of tokens to a little-endian u64 raw Amount, using the given mint. In this version of the program, the mint can only specify the number of decimals.

Return data can be fetched using sol_get_return_data and deserializing the return data as a little-endian u64.

Accounts expected by this instruction:

[] The mint to calculate for

Implementations

impl<'a> TokenInstruction<'a>

pub fn unpack(input: &'a [u8]) -> Result<Self, ProgramError>

Unpacks a byte buffer into a TokenInstruction.

pub fn pack(&self) -> Vec<u8>

Packs a TokenInstruction into a byte buffer.