spl_token
Module state

State transition types

Structs

Account

Account data.

Mint

Mint data.

Multisig

Multisignature data.

Enums

AccountState

Account state.

Constants

ACCOUNT_INITIALIZED_INDEX

The offset of state field in Account’s C representation

Traits

GenericTokenAccount

A trait for token Account structs to enable efficiently unpacking various fields without unpacking the complete state.

Functions

is_initialized_account

Check if the account data buffer represents an initialized account. This is checking the state (AccountState) field of an Account object.